#144 refactor: user 스키마를 mypage/follow 도메인으로 분리
